Make exif parser more robust
Currently, as soon as it finds a bad field, it gives up.
Some android devices consistently create a bad field (that
we dont care about) right before fields that we do care about
(the orientation, in particular).

Most exif parsers just move on to the next field in this case;
this changes ours to do the same.