disabled tests that only test underlying library
I'd rather not do this, but can't think of another good way. These 4 tests are failing on our opensource build since the image librarys are different there. The files they made end up looking the same but the bytes are different.

I just recently re-enabled these tests as they were a big comment (which I assumed was by accident) in the PHP-in-C++ version. (I'm using that to justify that we weren't covered before).