Fix uses of FBMAKE_BIN_ROOT
Its supposed to be a relative path from the fbcode dir,
but various tools assumed it was an absolute path.

In particular, test/run failed if it was set, but was
not an absolute path, and tools/run_test_binary.sh failed
if it was set and was an absolute path.