move repo contention test
I'd like to run the slow tests in our open source testing framework. This is the only test that fails hard. It actually fatals the test framework even with the 10 Meg limit we are using now.

What do you think about breaking it out to not run in open source?